very much working on refactoring, video only pops up on screen when video button is pressed
