+++ src/components/Cable.js
@@ -13,6 +13,11 @@ const Cable = ({ conversations, handleReceivedMessage, handleReceivedVideo }) =>
             channel={{ channel: 'VideosChannel', conversation: conversation.id }}
             onReceived={handleReceivedVideo}
           />
+          <ActionCable
+            key={"m" + conversation.id}
+            channel={{ channel: 'MessagesChannels', conversation: conversation.id }}
+            onReceived={handleReceivedMessage}
+          />
           </Fragment>
         );
       })}
